+++ vec3.hh
@@ -7,7 +7,7 @@
 #include "debug.hh"
 
 struct Vec3 {
-    Vec3(double x_, double y_, double z_):
+    Vec3(float x_, float y_, float z_):
         x(x_), y(y_), z(z_) {}
 
     Vec3& operator +=(const Vec3& b) {
@@ -24,14 +24,14 @@ struct Vec3 {
         return *this;
     }
 
-    Vec3& operator /=(const double b) {
+    Vec3& operator /=(const float b) {
         x /= b;
         y /= b;
         z /= b;
         return *this;
     }
 
-    Vec3 operator /(const double b) const {
+    Vec3 operator /(const float b) const {
         return Vec3(x / b,
                     y / b,
                     z / b);
@@ -49,7 +49,7 @@ struct Vec3 {
                     z - b.z);
     }
 
-    Vec3 operator *(const double b) const {
+    Vec3 operator *(const float b) const {
         return Vec3(b * x, b * y, b * z);
     }
 
@@ -60,11 +60,11 @@ struct Vec3 {
                     x * b.y - y * b.x);
     }
 
-    double dot(const Vec3& b) const {
+    float dot(const Vec3& b) const {
         return x * b.x + y * b.y + z * b.z;
     }
 
-    Vec3& operator *=(const double b) {
+    Vec3& operator *=(const float b) {
         x *= b;
         y *= b;
         z *= b;
@@ -82,11 +82,11 @@ struct Vec3 {
         return debug;
     }
 
-    double length() const {
+    float length() const {
         return sqrt(pow(x, 2) + pow(y, 2) + pow(z, 2));
     }
 
-    double distanceToPlane(const Vec3& normal, const double d) {
+    float distanceToPlane(const Vec3& normal, const float d) {
         return normal.x * x + normal.y * y + normal.z * z + d;
     }
 
@@ -97,9 +97,9 @@ struct Vec3 {
         return *this;
     }
 
-    double x;
-    double y;
-    double z;
+    float x;
+    float y;
+    float z;
 };
 
 #endif
